refactor: move App and misc files to new directories
updates directories for App component and misc styling / utility files
