+++ client/src/App/App.css
@@ -1,0 +1,209 @@
+/* dark mode */
+/* #root {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
+  transition: background-color 1s, color 1s; 
+} */
+
+/* .logo {
+  height: 2.5em;
+  padding: 1.5em;
+  filter: brightness(0) invert(1);
+  transition: filter 300ms;
+} */
+
+/* body.darkmode .logo {
+  filter: brightness(0) invert(1);
+}
+
+.logo:hover {
+  filter: drop-shadow(0 0 2em #646cffaa);
+}
+
+.logo.react:hover {
+  filter: drop-shadow(0 0 2em #61dafbaa);
+} */
+
+/* @keyframes logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+} */
+
+/* @media (prefers-reduced-motion: no-preference) {
+  a:nth-of-type(2) .logo {
+    animation: logo-spin infinite 20s linear;
+  }
+} */
+
+.card {
+  padding: 2em;
+}
+
+.read-the-docs {
+  color: #888;
+}
+
+/* .navbar {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 50px;
+  background-color: rgb(188, 188, 188);
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  z-index: 1000;
+  transition: background-color 0.3s;
+} */
+
+/* ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+}
+
+#navbarLinks {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0 34px;
+} */
+
+/* Dark mode styles */
+body.darkmode {
+  background-color: #1e1e1e;
+  color: #e0e0e0;
+}
+
+body.darkmode .navbar {
+  background-color: #333;
+}
+
+body.darkmode #navbarLinks a {
+  color: #e0e0e0;
+}
+
+body.darkmode .config-page-container {
+  background-color: #1e1e1e;
+  color: #e0e0e0;
+}
+
+/* chat style */
+
+.chat-container {
+  border: 1px solid #ddd;
+  background-color: #e1e1e1;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  border-radius: 8px;
+  padding: 16px;
+  margin-top: 20px;
+  max-height: 500px;
+  display: flex;
+  flex-direction: column;
+}
+
+.chat-title {
+  text-align: center;
+  font-size: 18px;
+  font-weight: bold;
+  margin-bottom: 16px;
+}
+
+.chat-window {
+  margin-bottom: 10px;
+  max-height: 300px;
+  overflow-y: auto;
+}
+
+.chat-message {
+  margin-bottom: 8px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.chat-message .role {
+  font-weight: bold;
+  margin-bottom: 4px;
+}
+
+.chat-input {
+  display: flex;
+  gap: 10px;
+  align-items: center;
+}
+
+.chat-input input {
+  flex: 1;
+  padding: 8px;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+}
+
+.chat-input button {
+  padding: 8px 16px;
+  border: none;
+  background-color: #447a90;
+  color: #ffffff;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.chat-input button:hover {
+  background-color: #62accc;
+}
+
+.chat-message p {
+  margin: 0;
+  padding: 5px;
+  border-radius: 4px;
+}
+
+.chat-message.user p {
+  background-color: #d1e7dd;
+}
+
+.chat-message.assistant p {
+  background-color: #e2e3e5;
+}
+
+body.darkmode .chat-container {
+  background-color: #2b2b2b;
+  color: #e0e0e0;
+  border-color: #444;
+}
+
+body.darkmode .chat-message {
+  background-color: #444;
+  color: #e0e0e0;
+}
+
+body.darkmode .chat-message.user p {
+  background-color: #444;
+}
+
+body.darkmode .chat-message.user {
+  background-color: #2b2b2b;
+  color: #e0e0e0;
+}
+
+body.darkmode .chat-message.assistant p {
+  background-color: #444;
+}
+body.darkmode .chat-message.assistant {
+  background-color: #2b2b2b;
+  color: #e0e0e0;
+}
+
+body.darkmode .chat-title {
+  color: #a2a2a2;
+}
