+++ client/src/App/App.tsx
@@ -1,11 +1,11 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import DashboardContainer from "./containers/FunctionPerformance//DashboardContainer";
-import CloudwatchContainer from "./containers/CloudWatchMetrics/CloudwatchContainer";
-import NavbarComponent from "./components/Navbar/NavbarComponent";
-import ConfigPageContainer from "./containers/ConfigPage/ConfigPageContainer";
-import ChatContainer from "./containers/FunctionPerformance/Chat/ChatContainer";
+import DashboardContainer from "../containers/FunctionPerformance//DashboardContainer";
+import CloudwatchContainer from "../containers/CloudWatchMetrics/CloudwatchContainer";
+import NavbarComponent from "../components/Navbar/NavbarComponent";
+import ConfigPageContainer from "../containers/ConfigPage/ConfigPageContainer";
+import ChatContainer from "../containers/FunctionPerformance/Chat/ChatContainer";
 import "./App.css";
-import "./chartSetup";
+import "../utils/chartSetup";
 
 function App() {
   return (
