+++ client/src/components/Navbar/NavbarComponent.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import "../../NavbarComponent.css";
+import "./Navbar.css";
 import lambda from "../../assets/lambda.png";
 
 const LightModeIcon = () => (
