+++ client/src/components/Row/RowComponent.tsx
@@ -1,4 +1,4 @@
-import "../../Graphs.css";
+import "../../styles/Graphs.css";
 
 const RowComponent = ({
   functionName,
