+++ client/src/containers/FunctionPerformance/Chat/ChatContainer.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import "../../../App.css";
+import "../../../App/App.css";
 
 const ChatContainer = () => {
   const [messages, setMessages] = useState<
