+++ client/src/main.tsx
@@ -1,11 +1,10 @@
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App.js';
-import './App.scss';
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import App from "./App/App";
+import "./App/App.scss";
 
-
-createRoot(document.getElementById('root')!).render(
+createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <App />
-  </StrictMode>,
+  </StrictMode>
 );
